backend/requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.1
PyPDF2==3.0.1
openai==1.72.0
pandas
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2